+++ patches/tModLoader/Terraria/ID/ItemRarityID.cs
@@ -1,11 +1,9 @@
 using ReLogic.Reflection;
-using System;
-using Terraria.ModLoader;
 
 namespace Terraria.ID
 {
 	/// <summary>Enumerates the values used with Item.rare</summary>
-	public class ItemRarityID
+	public static class ItemRarityID
 	{
 		/// <summary>Minus thirteen (-13)<para/>Master: Fiery Red<para/>Flag: item.master</summary>
 		public const int Master = -13;
@@ -40,35 +38,6 @@ namespace Terraria.ID
 		/// <summary>Eleven (11)</summary>
 		public const int Purple = 11;
 		public const int Count = 12;
-		public static readonly IdDictionary Search = IdDictionary.Create<ItemRarityID, int>();
-
-		public static string GetUniqueKey(int type) {
-			if (type < Master || (type > Quest && type < Gray) || type > RarityLoader.RarityCount)
-				throw new ArgumentOutOfRangeException("Invalid type: " + type);
-
-			if (type < Count)
-				return "Terraria " + Search.GetName(type);
-
-			var modRarity = RarityLoader.GetRarity(type);
-			return $"{modRarity.Mod.Name} {modRarity.Name}";
-		}
-
-		public static int TypeFromUniqueKey(string key) {
-			string[] parts = key.Split(new char[] { ' ' }, 2);
-			if (parts.Length != 2)
-				return 0;
-
-			return TypeFromUniqueKey(parts[0], parts[1]);
-		}
-		public static int TypeFromUniqueKey(string mod, string name) {
-			if (mod == "Terraria") {
-				if (!Search.ContainsName(name))
-					return 0;
-
-				return Search.GetId(name);
-			}
-
-			return ModLoader.ModLoader.GetMod(mod)?.RarityType(name) ?? 0;
-		}
+		public static readonly IdDictionary Search = IdDictionary.Create(typeof(ItemRarityID), typeof(int));
 	}
 }
