+++ patches/tModLoader/Terraria/ModLoader/Colors_tML.cs
@@ -1,0 +1,10 @@
+﻿using Microsoft.Xna.Framework;
+
+namespace Terraria.ID
+{
+	public static partial class Colors
+	{
+		public static readonly Color RarityDarkRed = new Color(255, 40, 100);
+		public static readonly Color RarityDarkPurple = new Color(180, 40, 255);
+	}
+}
